enable creating read-only subvolume
Because this feature is subvolume specific avoid creating a generic
config option and instead add an additional output format:

OutputFormat.subvolume_ro

To avoid having to test for this and the existing OutputFormat.subvolume
in various parts of the code add state.output_format which is the same as
config.output_format EXCEPT in this case when it is set to
Output.subvolume.

When config. and state. output_format differ that indicates to set the
output subvolume read-only.

Resolves: https://github.com/systemd/mkosi/issues/1534